% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chradle.R
\name{chr_call}
\alias{chr_call}
\title{chr_call}
\usage{
chr_call(instance, query)
}
\arguments{
\item{instance}{a chromium session returned by chr_init}

\item{query}{a message to be sent to the chromium session}
}
\value{
The chromium debugger response to the query.
}
\description{
Make a call to a chadle
}
\author{
Miles McBain
}
