% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chradle.R
\name{chr_clean}
\alias{chr_clean}
\title{chr_clean}
\usage{
chr_clean()
}
\value{
nothing
}
\description{
Clean up chromium process temp files
}
\details{
The background process will create a bunch of temporary files in the current
working directory under the folder 'remote-profile'. This will delete them.
}
\author{
Miles McBain
}
