% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chradle.R
\name{chr_init}
\alias{chr_init}
\title{chr_init}
\usage{
chr_init(url = "http://localhost:8080", debug_port = 9222,
  bin = "chromium-browser",
  block_on_message_pattern = "Runtime.executionContextCreated")
}
\arguments{
\item{url}{url of the session to be inspected}

\item{debug_port}{port to open the chromium debugger on}

\item{bin}{name of the browser binary to use, assumed to be on PATH.}

\item{block_on_message_pattern}{a text pattern appearing in a message to
wait for. Initialisation will not complete until this message is received.
Initialisation will fail if it is not received in around 30 seconds.}
}
\value{
a chromium session as a processx background process
}
\description{
Initialise a chradle
}
