% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chradle.R
\name{chr_kill}
\alias{chr_kill}
\title{chr_kill}
\usage{
chr_kill(instance)
}
\arguments{
\item{instance}{a chromium process running in background. Created with chr_init}
}
\value{
a response code for kill signal.
}
\description{
Kill a background chromium process created with chr_init()
}
