% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chradle.R
\name{get_attribute_names}
\alias{get_attribute_names}
\title{get_attribute_names}
\usage{
get_attribute_names(id)
}
\arguments{
\item{id}{of the html element to get attribute names}
}
\value{
A message to be sent with chr_call().
}
\description{
Compose a gettAttributeNames() call to be invoked in a chromium session.
}
