% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chradle.R
\name{get_attributes}
\alias{get_attributes}
\title{get_attributes()}
\usage{
get_attributes(id, attribute)
}
\arguments{
\item{id}{the id of the html element to get an attribute value from.}

\item{attribute}{the name of the attribute to get the value of.}
}
\value{
a message to be sent with chr_call().
}
\description{
Compose a gettAttribute() call to be invoked in a chromium session.
}
