% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chradle.R
\name{chr_init}
\alias{chr_init}
\title{chr_init}
\usage{
chr_init(url = "http://localhost:8080", debug_port = 9222,
  bin = "chromium-browser")
}
\arguments{
\item{url}{url of the session to be inspected}

\item{debug_port}{port to open the chromium debugger on}

\item{bin}{name of the browser binary to use, assumed to be on PATH.}
}
\value{
a chromium session as a processx background process
}
\description{
Initialise a chradle
}
